requirements.txt
streamlit
google-generativeai
pypdf2
pandas